build and run solvespace
* cannot use submodule for solvespace source as cmake fails

CMAKE ERROR
==========

Attempting to build solvespace when solvespace is a git submodule
fails b/c cmake is trying to get the git hash, and there is no
`.git` subdirectory in a submodule. As a workaround, I'm just manually
cloning the solvespace sources into this directory on build. Gross,
but it works.